Initial commit
-Mobile control implementation (touch-capable buttons with custom SVG icons)

-Rearanged file structure (grouped similar files in their own directories)

-Some cleanup done to keyboard control logic
