+++ css/style.sass
@@ -1,0 +1,83 @@
+// Overflow hidden keeps page from scrolling if spacebar is hit
+body 
+    margin: 0
+    overflow: hidden
+
+    // Disables pinch & zoom on mobile
+    touch-action: none
+
+// Added dynamic sizing to the canvas to make it scale to the users viewport
+// touch-action applied to disable "double-tap zoom" on mobile
+canvas 
+    border: 1px solid black
+    background: black
+    width: 100%
+    height: 100vh
+    touch-action: manipulation
+
+// Styling for mobile buttons
+#mobile-controls
+    display: none
+.mobile-btn
+    position: absolute
+    width: 70px
+    height: 70px
+    border-radius: 50%
+    z-index: 3
+    user-select: none
+    font-weight: 700
+    text-shadow: 2px 2px #111
+    background: transparent
+    opacity: .8
+    &:focus
+        outline: none
+    &#accel
+        background: url("/icons/burn.svg") no-repeat center center
+        background-size: 50%
+        bottom: 10%
+        left: 10%
+        color: #00adef
+        border-color: #00adef
+    &#fire
+        background: url("/icons/rocket.svg") no-repeat center center
+        background-size: 40%
+        bottom: 25%
+        left: 10%
+        color: #ed1c24
+        border-color: #ed1c24
+    &#left
+        background: url("/icons/ccw.svg") no-repeat center center
+        background-size: 40%
+        bottom: 15%
+        right: 30%
+        color: #3add5e
+        border-color: #3add5e
+    &#right
+        background: url("/icons/cw.svg") no-repeat center center
+        background-size: 40%
+        bottom: 15%
+        right: 7%
+        color: #3add5e
+        border-color: #3add5e
+
+    @media screen and ( min-width: 576px)
+        &#accel
+            bottom: 10%
+            left: 10%
+        &#fire
+            bottom: 35%
+            left: 10%
+        &#left
+            bottom: 15%
+            right: 30%
+        &#right
+            bottom: 15%
+            right: 10%
+#touch-toggle
+    display: none
+    top: 10%
+    right: 5%
+    color: white
+    border-radius: 15%
+    opacity: .5
+
