+++ spacewar.html
@@ -2,19 +2,32 @@
 <html lang="en">
 <head>
 	<meta charset="UTF-8">
+
+	<!-- added mobile-friendly meta data  -->
+	<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
+	<meta http-equiv="X-UA-Compatible" content="ie=edge">
+
 	<title>Spacewar!</title>
-	<style type="text/css">
-		canvas {
-			border: 1px solid black;
-			background: black;
-		}
-		body {
-			margin: 0;
-		}
-	</style>
+
+	<!-- moved styling to separate css file -->
+	<link rel="stylesheet" href="css/style.css">
+
 </head>
 <body>
 	<canvas></canvas>
-	<script src="spacewar.js" type="module"></script>
+
+	<!-- Start Mobile Controls -->
+	<div id="mobile-controls">
+		<button id="accel" class="mobile-btn"></button>
+		<button id="fire" class="mobile-btn"></button>
+		<button id="left" class="mobile-btn"></button>
+		<button id="right" class="mobile-btn"></button>
+	</div>
+	<button id="touch-toggle" class="mobile-btn">Toggle<br>Controls</button>
+	<!-- End Mobile Controls -->
+
+	<script src="js/spacewar.js" type="module"></script>
 </body>
 </html>
+
+<!-- Modified by Jeff Puls 12/22/2019 -->
